refactor: remove backend deps from preferences helpers
