+++ shared-helpers/index.ts
@@ -7,6 +7,7 @@ export * from "./src/occupancyFormatting"
 export * from "./src/pdfs"
 export * from "./src/photos"
 export * from "./src/postmarkString"
+export * from "./src/preferences"
 export * from "./src/programHelpers"
 export * from "./src/stringFormatting"
 export * from "./src/unitTypes"
