+++ sites/partners/src/applications/PaperApplicationForm/sections/FormAlternateContact.tsx
@@ -141,7 +141,6 @@ const FormAlternateContact = () => {
         <FormAddress
           subtitle={t("application.contact.mailingAddress")}
           dataKey="application.alternateContact.mailingAddress"
-          type="alternate"
           register={register}
           stateKeys={stateKeys}
         />
