+++ sites/partners/src/applications/PaperApplicationForm/sections/FormPreferences.tsx
@@ -2,19 +2,21 @@ import React, { useMemo } from "react"
 import {
   Field,
   t,
-  ExtraField,
-  getPreferenceOptionName,
   GridSection,
   ViewItem,
   GridCell,
   SelectOption,
-  getExclusivePreferenceOptionName,
-  getExclusiveKeys,
-  setExclusive,
 } from "@bloom-housing/ui-components"
 
 import { useFormContext } from "react-hook-form"
-import { stateKeys } from "@bloom-housing/shared-helpers"
+import {
+  stateKeys,
+  setExclusive,
+  getExclusiveKeys,
+  getExclusivePreferenceOptionName,
+  getPreferenceOptionName,
+  ExtraField,
+} from "@bloom-housing/shared-helpers"
 import {
   Preference,
   FormMetadataExtraData,
