+++ sites/public/pages/applications/preferences/all.tsx
@@ -6,17 +6,10 @@ import {
   FormCard,
   Field,
   t,
-  ExtraField,
   ExpandableContent,
   Button,
   AppearanceStyleType,
   resolveObject,
-  mapPreferencesToApi,
-  mapApiToPreferencesForm,
-  getPreferenceOptionName,
-  getExclusivePreferenceOptionName,
-  getExclusiveKeys,
-  setExclusive,
   ProgressNav,
   AuthContext,
 } from "@bloom-housing/ui-components"
@@ -24,7 +17,19 @@ import FormsLayout from "../../../layouts/forms"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 import { FormMetadataExtraData, Preference } from "@bloom-housing/backend-core/types"
-import { stateKeys, OnClientSide, PageView, pushGtmEvent } from "@bloom-housing/shared-helpers"
+import {
+  stateKeys,
+  OnClientSide,
+  PageView,
+  pushGtmEvent,
+  mapPreferencesToApi,
+  mapApiToPreferencesForm,
+  getPreferenceOptionName,
+  getExclusivePreferenceOptionName,
+  getExclusiveKeys,
+  setExclusive,
+  ExtraField,
+} from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../../../lib/constants"
 
 const ApplicationPreferencesAll = () => {
